Move CompileContiki into BaseContikiMoteType
This makes it clear that compilation logic is not
a dialog.
